Dice puzzle. CM Change, Moon, Animation Image, Shadows
Dice puzzle are now 3 dice instead of 5
changed the CM to change differently depending on the animation or in puzzle, moon is less brighter and the lamp as well.
added a spot light on the image in the animation start and disabled it mid animation.
added shadows in the corridor and the dream room